feat: service, implement ExcelReader with Apache POI

+++ backend/src/main/java/com/ffrontera/_generales/backend/modules/catalog/dto/PriceListRowDTO.java
@@ -1,8 +1,10 @@
 package com.ffrontera._generales.backend.modules.catalog.dto;
 
+import java.math.BigDecimal;
+
 public record PriceListRowDTO(
         String supplierSku,
         String Description,
-        Double newCost
+        BigDecimal newCost
 ) {
 }
